remove a ton of junk
