+++ app/layout.tsx
@@ -1,7 +1,6 @@
 import { CommandMenu } from '@/components/command';
 import { CommandMenuProvider } from '@/components/command/command-menu-context';
 import InteractiveConfigContextWrapper from '@/components/interactive/ContextWrapper';
-import Head from '@/components/idiot/appwrapper/Head';
 import { SidebarContentProvider } from '@/components/idiot/appwrapper/SidebarContentManager';
 import { SidebarContext } from '@/components/idiot/appwrapper/SidebarContext';
 import { SidebarMain } from '@/components/idiot/appwrapper/SidebarMain';
@@ -19,6 +18,25 @@ import { SolanaWalletProvider } from '@/components/wais/wallet/wallet-provider';
 const inter = Inter({ subsets: ['latin'] });
 
 export { metadata, viewport };
+export function Head() {
+  return (
+    <head>
+      <link rel='icon' href='/favicon.ico' sizes='any' />
+      <meta name='google-adsense-account' content={process.env.NEXT_PUBLIC_ADSENSE_ACCOUNT ?? ''} />
+      <meta property='og:url' content={process.env.NEXT_PUBLIC_APP_URI ?? ''} />
+      <meta property='og:type' content='website' />
+      <meta property='og:title' content={process.env.NEXT_PUBLIC_APP_NAME ?? 'NextJS Application'} />
+      <meta
+        property='og:description'
+        content={process.env.NEXT_PUBLIC_APP_DESCRIPTION ?? 'An application built with NextJS.'}
+      />
+      <meta
+        property='og:image'
+        content={process.env.NEXT_PUBLIC_APP_LOGO_URI || `${process.env.NEXT_PUBLIC_APP_URI}/favicon.ico`}
+      />
+    </head>
+  );
+}
 
 export default function RootLayout({ children }: { children: ReactNode }): ReactNode {
   const cookieStore = cookies();
